requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.1
django-cors-headers==3.13.0
django-on-heroku==1.1.2
djangorestframework==3.14.0
djangorestframework-filters==0.11.1
psycopg2-binary==2.9.3
python-dotenv==0.21.0
pytz==2022.2.1
sqlparse==0.4.3
whitenoise==6.2.0
